Fixed --list, check if args.list before running bust()
